Renamed UploadDialog, extracted UploadTabs. Updated UploadPage to have tabs.
