+++ webapp/src/Components/Pages/HomePage.tsx
@@ -26,7 +26,7 @@ import {getLoggedInUser, getReplayCount} from "../../Requests/Global"
 import {LinkButton} from "../Shared/LinkButton"
 import {Logo} from "../Shared/Logo/Logo"
 import {Search} from "../Shared/Search"
-import {UploadModalWrapper} from "../Shared/Upload/UploadModalWrapper"
+import {UploadDialogWrapper} from "../Shared/Upload/UploadDialogWrapper"
 
 type Props = ReturnType<typeof mapStateToProps>
     & ReturnType<typeof mapDispatchToProps>
@@ -57,7 +57,7 @@ class HomePageComponent extends React.PureComponent<Props, State> {
         const alignCenterProps: GridProps = {container: true, justify: "center", alignItems: "center"}
         return (
             <div className={classes.backgroundContainer}>
-                <UploadModalWrapper buttonStyle="floating">
+                <UploadDialogWrapper buttonStyle="floating">
                     <div className={classes.root}>
                         <Grid container justify="center" alignItems="flex-start" spacing={40} className={classes.child}>
                             <Grid item xs={12} {...alignCenterProps} style={{minHeight: "300px"}} direction="column">
@@ -106,7 +106,7 @@ class HomePageComponent extends React.PureComponent<Props, State> {
                             </Grid>
                         </Grid>
                     </div>
-                </UploadModalWrapper>
+                </UploadDialogWrapper>
             </div>
         )
     }
