+++ webapp/src/Components/Pages/UploadPage.tsx
@@ -1,18 +1,41 @@
-import {Grid} from "@material-ui/core"
+import {Card, Divider, Grid} from "@material-ui/core"
 import * as React from "react"
+import {PreviousUploads} from "../Shared/Upload/PreviousUploads"
 import {UploadForm} from "../Shared/Upload/UploadForm"
+import {UploadTab, UploadTabs} from "../Shared/Upload/UploadTabs"
 import {BasePage} from "./BasePage"
 
-export class UploadPage extends React.PureComponent {
+interface State {
+    selectedTab: UploadTab
+}
+
+export class UploadPage extends React.PureComponent<{}, State> {
+    constructor(props: {}) {
+        super(props)
+        this.state = {selectedTab: "Upload Replays"}
+    }
+
     public render() {
         return (
             <BasePage>
                 <Grid container spacing={16} alignItems="center" justify="center">
                     <Grid item>
-                        <UploadForm/>
+                        <Card>
+                            <UploadTabs selectedTab={this.state.selectedTab} handleChange={this.handleTabChange}/>
+                            <Divider/>
+                            {this.state.selectedTab === "Upload Replays" ?
+                                <UploadForm/>
+                                :
+                                <PreviousUploads/>
+                            }
+                        </Card>
                     </Grid>
                 </Grid>
             </BasePage>
         )
+    }
+
+    private readonly handleTabChange = (event: React.ChangeEvent, selectedTab: UploadTab) => {
+        this.setState({selectedTab})
     }
 }
