+++ webapp/src/Components/Shared/NavBar/NavBar.tsx
@@ -23,7 +23,7 @@ import {setLoggedInUserAction} from "../../../Redux/loggedInUser/actions"
 import {getLoggedInUser} from "../../../Requests/Global"
 import {Logo} from "../Logo/Logo"
 import {Search} from "../Search"
-import {UploadModalWrapper} from "../Upload/UploadModalWrapper"
+import {UploadDialogWrapper} from "../Upload/UploadDialogWrapper"
 import {AccountMenu} from "./AccountMenu"
 
 type Props = ReturnType<typeof mapStateToProps>
@@ -72,7 +72,7 @@ class NavBarComponent extends React.PureComponent<Props> {
                                 </Link>
                             </Grid>
                             <Grid item>
-                                <UploadModalWrapper buttonStyle="icon"/>
+                                <UploadDialogWrapper buttonStyle="icon"/>
                             </Grid>
                             <Grid item xs="auto" className={classes.accountMenuGridItem}>
                                 {/*TODO: Show AccountMenu even on mobile, but as a 3 dots icon.*/}
